Add NoDataLoss param

+++ functions/Install-DbaMaintenanceSolution.ps1
@@ -44,6 +44,12 @@ function Install-DbaMaintenanceSolution {
         Specifies the path to a local file to install Ola's solution from. This *should* be the zip file as distributed by the maintainers.
         If this parameter is not specified, the latest version will be downloaded and installed from https://github.com/olahallengren/sql-server-maintenance-solution
 
+    .PARAMETER NoDataLoss
+        Normally, when the ReplaceExisting flag is used, it will drop the
+        CommandLog table and re-create it. When this flag is also passed, it
+        will instead rename the old table with a time stamp which will allow
+        you to do a migration of existing data to the new table.
+
     .PARAMETER Force
         If this switch is enabled, the Ola's solution will be downloaded from the internet even if previously cached.
 
@@ -138,6 +144,7 @@ function Install-DbaMaintenanceSolution {
         [string[]]$Solution = 'All',
         [switch]$InstallJobs,
         [string]$LocalFile,
+        [switch]$NoDataLoss,
         [switch]$Force,
         [switch]$EnableException
     )
@@ -344,9 +351,22 @@ function Install-DbaMaintenanceSolution {
 
             $CleanupQuery = $null
             if ($ReplaceExisting) {
-                [string]$CleanupQuery = $("
+
+                if ($NoDataLoss -eq $true) {
+                    [string]$CleanupQuery = $("
+                            IF OBJECT_ID('[dbo].[CommandLog]', 'U') IS NOT NULL
+                                exec sp_rename '[dbo].[CommandLog]', 'CommandLog_{0:yyyy-MM-dd}T{0:HH:mm:ss}Z';
+                            IF OBJECT_ID('[dbo].[PK_CommandLog]', 'U') IS NOT NULL
+                                exec sp_rename 'PK_CommandLog', 'PK_CommandLog_{0:yyyy-MM-dd}T{0:HH:mm:ss}Z'
+                    ") -f ((get-date).ToUniversalTime());
+                } else {
+                    [string]$CleanupQuery = $("
                             IF OBJECT_ID('[dbo].[CommandLog]', 'U') IS NOT NULL
                                 DROP TABLE [dbo].[CommandLog];
+                ");
+                }
+
+                $CleanupQuery += $("
                             IF OBJECT_ID('[dbo].[CommandExecute]', 'P') IS NOT NULL
                                 DROP PROCEDURE [dbo].[CommandExecute];
                             IF OBJECT_ID('[dbo].[DatabaseBackup]', 'P') IS NOT NULL
@@ -356,7 +376,7 @@ function Install-DbaMaintenanceSolution {
                             IF OBJECT_ID('[dbo].[IndexOptimize]', 'P') IS NOT NULL
                                 DROP PROCEDURE [dbo].[IndexOptimize];
                             ")
-
+                write-host -message $CleanupQuery;
                 if ($Pscmdlet.ShouldProcess($instance, "Dropping all objects created by Ola's Maintenance Solution")) {
                     Write-ProgressHelper -ExcludePercent -Message "Dropping objects created by Ola's Maintenance Solution"
                     $null = $db.Query($CleanupQuery)
